LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-from-initial-about-blank-src-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL
CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Invalid URL


FAIL navigate() from <iframe> with src="" but still on initial about:blank works assert_equals: expected "/common/blank.html" but got "/navigation-api/navigation-methods/navigate-from-initial-about-blank-src.html"
